+++ lib/QuantumToolboxCore/Project.toml
@@ -4,6 +4,7 @@ authors = ["Alberto Mercurio", "Yi-Te Huang"]
 version = "0.1.0"
 
 [deps]
+FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SciMLOperators = "c0aeaf25-5076-4817-a8d5-81caf7dfa961"
@@ -11,6 +12,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArraysCore = "1e83bf80-4336-4d27-bf5d-d5a4f845583c"
 
 [compat]
+FillArrays = "1"
 LinearAlgebra = "1"
 Random = "1"
 SciMLOperators = "1.12"
