+++ lib/QuantumToolboxCore/src/QuantumToolboxCore.jl
@@ -1,11 +1,40 @@
 module QuantumToolboxCore
 
-using LinearAlgebra
+# standard Julia libraries
 using SparseArrays
-import StaticArraysCore: SVector, MVector
-import Random: AbstractRNG
+using LinearAlgebra
+import LinearAlgebra: checksquare
 
-import SciMLOperators: SciMLOperators, AbstractSciMLOperator
+# other dependencies (in alphabetical order)
+import Base: AbstractVecOrTuple
+import FillArrays: Eye
+import Random: AbstractRNG, default_rng
+import StaticArraysCore: SVector, MVector
+
+import SciMLOperators:
+    SciMLOperators,
+    cache_operator,
+    iscached,
+    isconstant,
+    AbstractSciMLOperator,
+    MatrixOperator,
+    ScalarOperator,
+    ScaledOperator,
+    AddedOperator,
+    ComposedOperator,
+    update_coefficients!,
+    concretize
+
+# Export functions from the other modules
+
+## LinearAlgebra
+export ishermitian, issymmetric, isposdef, dot, tr, svdvals, norm, normalize, normalize!, diag, Hermitian, Symmetric
+
+## SparseArrays
+export permute
+
+## SciMLOperators
+export cache_operator, iscached, isconstant
 
 ## Basic utilities for QuantumToolbox libraries
 include("settings.jl")
@@ -20,6 +49,26 @@ include("linalg/linear_maps.jl")
 ## Other physical related utilities
 include("PhysicalConstants.jl")
 include("physics_func.jl")
+
+## Quantum Object
+include("qobj/dimensions.jl")
+include("qobj/energy_restricted.jl")
+include("qobj/quantum_object_base.jl")
+include("qobj/quantum_object.jl")
+include("qobj/quantum_object_evo.jl")
+include("qobj/boolean_functions.jl")
+include("qobj/arithmetic_and_attributes.jl")
+include("qobj/eigen.jl")
+include("qobj/functions.jl")
+include("qobj/states.jl")
+include("qobj/operators.jl")
+include("qobj/superoperators.jl")
+include("qobj/synonyms.jl")
+
+## Other functionalities for QuantumObject
+include("entropy.jl")
+include("metrics.jl")
+include("negativity.jl")
 
 ## deprecated functions
 include("deprecated.jl")
