+++ lib/QuantumToolboxCore/src/qobj/energy_restricted.jl
