+++ lib/QuantumToolboxCore/src/qobj/synonyms.jl
