+++ src/block_diagonal_form.jl
